+++ ddot.h
@@ -1,4 +1,4 @@
 #ifndef DDOT_H
 #define DDOT_H
-int ddot (const int n, const double * const x, const double * const y, double * const result);
+int ddot (const int n, const float * const x, const float * const y, float * const result);
 #endif
