+++ generate_matrix.h
@@ -3,5 +3,5 @@
 
 #include "mesh.h"
 
-void generate_matrix(int nx, int ny, int nz, struct mesh **A, double **x, double **b, double **xexact, int use_7pt_stencil);
+void generate_matrix(int nx, int ny, int nz, struct mesh **A, float **x, float **b, float **xexact, int use_7pt_stencil);
 #endif
